# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: OntoConsensus API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /projects:
    get:
      operationId: listProjects
      tags: [projects]
      summary: List projects the current user belongs to
      responses:
        "200":
          description: Projects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ProjectSummary"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      operationId: createProject
      tags: [projects]
      summary: Create a project from an uploaded ontology file
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/ProjectInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectSummary"
        "400":
          description: Invalid file or missing classes
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /projects/join:
    post:
      operationId: joinProject
      tags: [projects]
      summary: Join a project using an invite code
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JoinProjectInput"
      responses:
        "200":
          description: Joined
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectSummary"
        "400":
          description: Invalid code or project full
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /projects/{id}:
    get:
      operationId: getProject
      tags: [projects]
      summary: Get project details, members and graph
      parameters:
        - $ref: "#/components/parameters/ProjectId"
      responses:
        "200":
          description: Project detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectDetail"
        "403":
          description: Not a member
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteProject
      tags: [projects]
      summary: Delete a project (owner only) — removes it for every member
      parameters:
        - $ref: "#/components/parameters/ProjectId"
      responses:
        "204":
          description: Deleted
        "403":
          description: Only the owner can delete this project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /projects/{id}/ws-ticket:
    post:
      operationId: createWsTicket
      tags: [projects]
      summary: Create a short-lived ticket to open the realtime connection
      parameters:
        - $ref: "#/components/parameters/ProjectId"
      responses:
        "200":
          description: Ticket
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WsTicket"
...
  /projects/{id}/export:
    get:
      operationId: exportProject
      tags: [projects]
      summary: Export classes and fully-agreed properties as JSON
      parameters:
        - $ref: "#/components/parameters/ProjectId"
      responses:
        "200":
          description: Export file
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExportPayload"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  parameters:
    ProjectId:
      name: id
      in: path
      required: true
      schema:
        type: integer
...
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    ProjectInput:
      type: object
      required: [name, file, memberCount]
      properties:
        name:
          type: string
          minLength: 1
        file:
          type: string
          format: binary
        memberCount:
          type: integer
          minimum: 1
          maximum: 3
          description: >-
            Exact number of members this project is for, chosen once at
            creation and fixed afterward. Invites are capped at this number,
            the shared consensus space only opens once exactly this many
            members have all marked ready, and each member's private
            property budget is derived from this number from the start.
        moderatorEnabled:
          type: boolean
          default: true
          description: >-
            Whether the AI moderator's messages are shown in the live
            conversation window for this project, chosen once at creation and
            fixed afterward -- there is no route to change it later. The
            moderator itself always keeps running in the background
            regardless of this setting (it still listens, transcribes, and
            generates interventions); disabling this only hides its "intro"
            and "intervention" messages from the live chat UI. Its messages
            still appear normally, in their original position, in the
            conversation export. Defaults to true.
    JoinProjectInput:
      type: object
      required: [inviteCode]
      properties:
        inviteCode:
          type: string
...
    WsTicket:
      type: object
      required: [ticket]
      properties:
        ticket:
          type: string
...
    Member:
      type: object
      required: [userId, username, colorSlot, ready, isOwner]
      properties:
        userId:
          type: integer
        username:
          type: string
        colorSlot:
          type: integer
        ready:
          type: boolean
        isOwner:
          type: boolean
    OntologyClass:
      type: object
      required: [id, uri, label, propertyCount, atPropertyCap]
      properties:
        id:
          type: integer
        uri:
          type: string
        label:
          type: string
        propertyCount:
          type: integer
          description: >-
            Number of distinct properties the REQUESTING member has proposed
            for this class. Each member has their own fixed budget (see
            atPropertyCap) — there is no shared/global cap and no cross-member
            duplicate check, so this never reflects other members' proposals.
        atPropertyCap:
          type: boolean
          description: >-
            True once the requesting member has reached their own budget of
            distinct properties for this class (budget depends on project
            member count and join order — e.g. 7/4+3/3+2+2 for 1/2/3
            members). Used to hide the "add property" affordance once THIS
            member is out of budget; other members may still have room.
    OntologyRelation:
      type: object
      required: [childId, parentId]
      properties:
        childId:
          type: integer
        parentId:
          type: integer
    ProjectSummary:
      type: object
      required:
        [
          id,
          name,
          inviteCode,
          ownerId,
          memberCount,
          maxMembers,
          moderatorEnabled,
          createdAt,
        ]
      properties:
        id:
          type: integer
        name:
          type: string
        inviteCode:
          type: string
        ownerId:
          type: integer
        memberCount:
          type: integer
        maxMembers:
          type: integer
        moderatorEnabled:
          type: boolean
          description: >-
            Whether the AI moderator's messages are shown in this project's
            live conversation window. Fixed at creation.
        exportedAt:
          type: string
          format: date-time
          nullable: true
          description: >-
            Set the moment any member first downloads the ontology export
            (see GET /projects/{id}/export). Once set, the AI moderator never
            runs again for this project and the conversation shown to
            members is frozen at the consensus reached at that point. Null
            means the project hasn't been exported yet.
        createdAt:
          type: string
          format: date-time
    ProjectDetail:
      type: object
      required:
        [
          id,
          name,
          inviteCode,
          ownerId,
          maxMembers,
          moderatorEnabled,
          createdAt,
          members,
          classes,
          relations,
        ]
      properties:
        id:
          type: integer
        name:
          type: string
        inviteCode:
          type: string
        ownerId:
          type: integer
        maxMembers:
          type: integer
        moderatorEnabled:
          type: boolean
          description: >-
            Whether the AI moderator's messages are shown in this project's
            live conversation window. Fixed at creation.
        exportedAt:
          type: string
          format: date-time
          nullable: true
          description: >-
            Set the moment any member first downloads the ontology export
            (see GET /projects/{id}/export). Once set, the AI moderator never
            runs again for this project and the conversation shown to
            members is frozen at the consensus reached at that point. Null
            means the project hasn't been exported yet.
        createdAt:
          type: string
          format: date-time
        members:
          type: array
          items:
            $ref: "#/components/schemas/Member"
        classes:
          type: array
          items:
            $ref: "#/components/schemas/OntologyClass"
        relations:
          type: array
          items:
            $ref: "#/components/schemas/OntologyRelation"
...
    ExportClass:
      type: object
      required: [label, properties, children]
      properties:
        label:
          type: string
        properties:
          type: array
          items:
            type: string
          description: Only properties every specified member has agreed on.
        children:
          type: array
          items:
            $ref: "#/components/schemas/ExportClass"
          description: Direct subclasses, nested recursively — represents the ontology's class hierarchy structurally instead of via a parent reference.
    ExportMeta:
      type: object
      required: [projectName, exportedAt, memberCount]
      properties:
        projectName:
          type: string
        exportedAt:
          type: string
          format: date-time
        memberCount:
          type: integer
          description: The project's specified number of members (not just however many have joined).
    ExportPayload:
      type: object
      required: [meta, classes]
      properties:
        meta:
          $ref: "#/components/schemas/ExportMeta"
        classes:
          type: array
          items:
            $ref: "#/components/schemas/ExportClass"
...