# lib/api-zod/src/generated/types/exportMeta.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * OntoConsensus API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ExportMeta {
  projectName: string;
  exportedAt: Date;
  /** The project's specified number of members (not just however many have joined). */
  memberCount: number;
}