# lib/api-zod/src/generated/types/projectDetail.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * OntoConsensus API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Member } from './member';
import type { OntologyClass } from './ontologyClass';
import type { OntologyRelation } from './ontologyRelation';

export interface ProjectDetail {
  id: number;
  name: string;
  inviteCode: string;
  ownerId: number;
  maxMembers: number;
  /** Whether the AI moderator's messages are shown in this project's live conversation window. Fixed at creation. */
  moderatorEnabled: boolean;
  /** Set the moment any member first downloads the ontology export (see GET /projects/{id}/export). Once set, the AI moderator never runs again for this project and the conversation shown to members is frozen at the consensus reached at that point. Null means the project hasn't been exported yet. */
  exportedAt?: Date | null;
  createdAt: Date;
  members: Member[];
  classes: OntologyClass[];
  relations: OntologyRelation[];
}